+++ angular/lib/src/compiler/view_compiler/expression_converter.dart
@@ -118,8 +118,7 @@ class _AstToIrVisitor implements compiler_ast.AstVisitor {
     this._implicitReceiver,
     this._preserveWhitespace,
     this._boundType,
-  )
-      : _visitingRoot = true {
+  ) : _visitingRoot = true {
     assert(_nameResolver != null);
   }