+++ angular_ast/lib/src/expression/micro/parser.dart
@@ -31,8 +31,7 @@ class NgMicroParser {
       expression.length,
       tokens,
       origin,
-    )
-        .parse();
+    ).parse();
   }
 }