+++ angular_compiler/lib/src/analyzer/di/providers.dart
@@ -205,8 +205,7 @@ class UseClassProviderElement extends ProviderElement {
     this.useClass, {
     @required this.dependencies,
     bool multi: false,
-  })
-      : super._(e, providerType, multi);
+  }) : super._(e, providerType, multi);
 
   @override
   bool operator ==(Object o) =>
@@ -238,8 +237,7 @@ class UseExistingProviderElement extends ProviderElement {
     TypeLink providerType,
     this.redirect, {
     bool multi: false,
-  })
-      : super._(e, providerType, multi);
+  }) : super._(e, providerType, multi);
 
   @override
   bool operator ==(Object o) =>
@@ -272,8 +270,7 @@ class UseFactoryProviderElement extends ProviderElement {
     this.useFactory, {
     @required this.dependencies,
     bool multi: false,
-  })
-      : super._(e, providerType, multi);
+  }) : super._(e, providerType, multi);
 
   @override
   bool operator ==(Object o) =>
@@ -307,6 +304,5 @@ class UseValueProviderElement extends ProviderElement {
     TypeLink providerType,
     this.useValue, {
     bool multi: false,
-  })
-      : super._(e, providerType, multi);
+  }) : super._(e, providerType, multi);
 }