+++ angular/CHANGELOG.md
@@ -14,6 +14,9 @@
 
 ### Bug Fixes
 
+*   Testability now includes ComponentState updates. Due to prior use of
+    animationFrame callback, testbed was not able to detect stable state.
+
 *   Misspelled or otherwise erroneous annotations on classes now produce a more
     understandable error message, including the element that was annotated and
     the annotation that was not resolved.